requirements.txt
asgiref==3.8.1
Django==4.2.13
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-nested-admin==4.0.2
django-nested-inline==0.4.6
django-widget-tweaks==1.5.0
packaging==24.0
pillow==10.3.0
plotly==5.22.0
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
typing_extensions==4.11.0
tzdata==2024.1
